refactor(describe): separate workload struct/funcs
