+++ internal/pkg/deploy/cloudformation/stack/name.go
@@ -24,11 +24,11 @@ func (t TaskStackName) TaskName() string {
 	return strings.SplitN(string(t), "-", 2)[1]
 }
 
-// NameForService returns the stack name for a service.
-func NameForService(app, env, svc string) string {
+// NameForWorkload returns the stack name for a workload.
+func NameForWorkload(app, env, name string) string {
 	// stack name limit constrained by CFN https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/cfn-using-console-create-stack-parameters.html
 	const maxLen = 128
-	stackName := fmt.Sprintf("%s-%s-%s", app, env, svc)
+	stackName := fmt.Sprintf("%s-%s-%s", app, env, name)
 
 	if len(stackName) > maxLen {
 		return stackName[:maxLen]
