+++ internal/pkg/deploy/cloudformation/stack/workload.go
@@ -147,7 +147,7 @@ type wkld struct {
 
 // StackName returns the name of the stack.
 func (w *wkld) StackName() string {
-	return NameForService(w.app, w.env, w.name)
+	return NameForWorkload(w.app, w.env, w.name)
 }
 
 // Parameters returns the list of CloudFormation parameters used by the template.
